fix some error

+++ z PHP install/nginx_include.sh
@@ -19,9 +19,12 @@ FILETEMP=/etc/nginx/certs
 		[[ ! -f $FILETEMP ]] || rm -f $FILETEMP
 	FILETEMP=/etc/nginx/sites-enabled/default
 		[[ ! -f $FILETEMP ]] || rm -f $FILETEMP
+		# download default config site
+		wget -O $FILETEMP --no-check-certificate $DOWN_URL/nginx_config/default.conf
 	FILETEMP=/etc/nginx/sites-enabled/default.conf
 		[[ ! -f $FILETEMP ]] || rm -f $FILETEMP
-		wget -O $FILETEMP --no-check-certificate $DOWN_URL/nginx_config/default.conf
+		# download default config site if no have
+		[[ -f /etc/nginx/sites-enabled/default ]] || wget -O $FILETEMP --no-check-certificate $DOWN_URL/nginx_config/default.conf
 	FILETEMP=/etc/nginx/http2-ssl.conf
 		[[ ! -f $FILETEMP ]] || rm -f $FILETEMP
 		wget -O $FILETEMP --no-check-certificate $DOWN_URL/nginx_config/http2-ssl.conf
